+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unpublished]
+### Fixed
+- Fixed typo on conda installation docs [#373]
+
 ## [4.6.1] - 2025-07-24
 ### Fixed
 - Publish to PyPI on release event action