+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## [unpublished]
+### Added
+- `email` and `additionalContacts` in API reference json schema ()
 ### Changed
-- Updated versions of external actions and MongoDB in GitHub workflows [#374]
+- Updated versions of external actions and MongoDB in GitHub workflows (#374)
 ### Fixed
-- Fixed typo on conda installation docs [#373]
+- Fixed typo on conda installation docs (#373)
 
 ## [4.6.1] - 2025-07-24
 ### Fixed