+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - `email`, `additionalContacts` and `roles` in API reference json schema (#380 and #384)
 - Allow assigning contact `roles` using the cli (#383)
+- A test to make sure 
 ### Changed
 - Updated versions of external actions and MongoDB in GitHub workflows (#374)
 - Allow updating contact's email and href separately using the command line (#381)