+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Conftest file to reflect a patient's contact with email, roles and additionalContacts (#383)
 ### Fixed
 - Fixed typo on conda installation docs (#373)
-- Parsing and saving `additionalContacts` field when a patient is saved using the endpoint ()
+- Parsing and saving `additionalContacts` field when a patient is saved using the endpoint (#386)
 
 ## [4.6.1] - 2025-07-24
 ### Fixed